Continue with refactoring
