+++ lua/config/keymaps.lua
@@ -5,17 +5,7 @@ keymap({ "n", "v" }, "<Space>", "<Nop>", { silent = true })
 vim.g.mapleader = " "
 vim.g.maplocalleader = ","
 
-local function ExCurrentDir()
-  local current_file_dir = vim.fn.expand('%:p:h')
-  if vim.bo.filetype == 'netrw' then
-    vim.cmd('bd!')
-  else
-    vim.cmd('Lex ' .. current_file_dir)
-  end
-end
-
-keymap("n", "<leader>ep", "<cmd>:Lex<CR>", opts)
-keymap('n', '<leader>et', ExCurrentDir, opts, "Project Tree")
+keymap("n", "<leader>ee", ":Ex<CR>", opts)
 
 keymap("n", "<C-i>", "<C-i>", opts)
