+++ lua/config/treesitter.lua
@@ -1,0 +1,34 @@
+  ---@diagnostic disable: missing-fields
+  require("nvim-treesitter.configs").setup({
+    ensure_installed = {
+      "lua",
+      "luadoc",
+      "vim",
+      "vimdoc",
+      "graphql",
+      "markdown",
+      "markdown_inline",
+      "bash",
+      "dockerfile",
+      "html",
+      "json",
+      "scss",
+      "xml",
+      "yaml",
+      "css",
+      "sql",
+      "c",
+      "odin",
+      "go",
+      "ocaml",
+      "python",
+      "typescript",
+      "javascript",
+      "clojure",
+      "rust",
+      "php"
+    },
+    highlight = { enable = true },
+    indent = { enable = true },
+    sync_install = false,
+  })
