+++ lua/plugins/mason.lua
@@ -5,33 +5,8 @@ local M = {
 	},
 }
 
-function M.config()
-	local servers = {
-		"lua_ls",
-		-- "emmet_language_server",
-		"cssls",
-		"htmx",
-		-- "html",
-		"emmet_ls",
-		"ts_ls",
-		"pyright",
-		"bashls",
-		"jsonls",
-		"clojure_lsp",
-		"intelephense",
-		"rust_analyzer",
-		"gopls",
-		"tailwindcss",
-		"graphql",
-		"ols",
-		"clangd"
-	}
-
-	require("mason").setup({})
-
-	require("mason-lspconfig").setup({
-		ensure_installed = servers,
-	})
+M.config = function ()
+	require("config.mason")
 end
 
 return M
