+++ lua/plugins/treesitter.lua
@@ -5,40 +5,7 @@ local M = {
 }
 
 function M.config()
-  ---@diagnostic disable: missing-fields
-  require("nvim-treesitter.configs").setup({
-    ensure_installed = {
-      "lua",
-      "luadoc",
-      "vim",
-      "vimdoc",
-      "graphql",
-      "markdown",
-      "markdown_inline",
-      "bash",
-      "dockerfile",
-      "html",
-      "json",
-      "scss",
-      "xml",
-      "yaml",
-      "css",
-      "sql",
-      "c",
-      "odin",
-      "go",
-      "ocaml",
-      "python",
-      "typescript",
-      "javascript",
-      "clojure",
-      "rust",
-      "php"
-    },
-    highlight = { enable = true },
-    indent = { enable = true },
-    sync_install = false,
-  })
+  require("config.treesitter")
 end
 
 return M
